feat: major code/layout refactoring
Moved all state management into a context provider. Cleaned up the subscription tracker layout by separating parts of the page into their own components. Made the root subscription tracker page a server side page.
