+++ app/context/ThemeContext.tsx
@@ -1,6 +1,7 @@
 "use client";
 
-import { createContext, useState, useEffect, ReactNode } from "react";
+import useLocalStorageState from "@/utils/useLocalStorageState";
+import { createContext, ReactNode } from "react";
 
 const DefaultTheme: string = "custom";
 
@@ -19,31 +20,13 @@ interface ThemeProviderProps {
 }
 
 export const ThemeProvider = ({ children }: ThemeProviderProps) => {
-  const [theme, setTheme] = useState<string>(DefaultTheme);
-  const [isMounted, setIsMounted] = useState<boolean>(false);
-
-  useEffect(() => {
-    if (typeof window !== "undefined") {
-      const storedTheme = localStorage.getItem("theme") || DefaultTheme;
-      setTheme(storedTheme);
-      setIsMounted(true);
-    }
-  }, []);
-
-  const changeTheme = (theme: string) => {
-    setTheme(theme);
-    if (typeof window !== "undefined") {
-      localStorage.setItem("theme", theme);
-    }
-  };
-
-  return isMounted ? (
+  const [theme, changeTheme] = useLocalStorageState<string>(
+    "theme",
+    DefaultTheme,
+  );
+  return (
     <ThemeContext.Provider value={{ theme, changeTheme }}>
       {children}
     </ThemeContext.Provider>
-  ) : (
-    <div className="flex h-screen items-center justify-center gap-4 bg-base-100 text-3xl font-bold uppercase">
-      <span className="loading loading-infinity loading-lg text-base-content" />
-    </div>
   );
 };
