+++ app/websites/subscription_tracker/context/SubscriptionTrackerContext.tsx
@@ -1,0 +1,373 @@
+"use client";
+
+import {
+  Dispatch,
+  FormEvent,
+  SetStateAction,
+  createContext,
+  useContext,
+  useState,
+} from "react";
+import Subscription from "../interfaces/Subscription";
+import {
+  compressJSON,
+  copyToClipboard,
+  decompressJSON,
+  formatCurrency,
+  formatDateToYyyyMmDd,
+  generateUUID,
+  isIndexOutOfRange,
+} from "@/utils/helperFunctions";
+import SubscriptionType from "../types/SubscriptionType";
+import useLocalStorageState from "@/utils/useLocalStorageState";
+
+const DefaultFormData: Subscription = {
+  id: "",
+  label: "",
+  price: 0,
+  type: "monthly",
+  firstPaymentDate: formatDateToYyyyMmDd(new Date()),
+};
+
+type Context = {
+  subscriptions: Subscription[];
+  setSubscriptions: Dispatch<SetStateAction<Subscription[]>>;
+  selectedSubscriptionIndex: number;
+  setSelectedSubscriptionIndex: Dispatch<SetStateAction<number>>;
+  formData: Subscription;
+  setFormData: Dispatch<SetStateAction<Subscription>>;
+  editingSubscription: boolean;
+  setEditingSubscription: Dispatch<SetStateAction<boolean>>;
+  locale: string;
+  setLocale: Dispatch<SetStateAction<string>>;
+  currency: string;
+  setCurrency: Dispatch<SetStateAction<string>>;
+  isImporting: boolean;
+  setIsImporting: Dispatch<SetStateAction<boolean>>;
+  importCode: string;
+  setImportCode: Dispatch<SetStateAction<string>>;
+  sortedSubscriptions: Subscription[];
+  addSubscription: (
+    label: string,
+    price: number,
+    type: SubscriptionType,
+    date: string,
+  ) => void;
+  calculateTotalPerMonth: () => number;
+  calculateTotalPerYear: () => number;
+  handleFormSubmit: (e: FormEvent) => void;
+  handleFormChange: (
+    e: React.ChangeEvent<
+      HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement
+    >,
+  ) => void;
+  editSubscription: (index: number) => void;
+  cancelEdit: () => void;
+  deleteSubscription: (index: number) => void;
+  getSubscriptionDueDate: (subscription: Subscription) => number;
+  isSubscriptionDueToday: (subscription: Subscription) => boolean;
+  getSubscriptionTotalSpend: (subscription: Subscription) => number;
+  formatToCurrencyString: (num: number) => string;
+  backupSubscriptions: () => Promise<void>;
+  importSubscriptions: () => void;
+  closeImport: () => void;
+  startImport: () => void;
+  deleteAllSubscriptions: () => void;
+};
+
+const SubscriptionTrackerContext = createContext<Context>({
+  subscriptions: [],
+  setSubscriptions: () => {},
+  selectedSubscriptionIndex: 0,
+  setSelectedSubscriptionIndex: () => {},
+  formData: DefaultFormData,
+  setFormData: () => {},
+  editingSubscription: false,
+  setEditingSubscription: () => {},
+  locale: "en-GB",
+  setLocale: () => {},
+  currency: "GBP",
+  setCurrency: () => {},
+  isImporting: false,
+  setIsImporting: () => {},
+  importCode: "",
+  setImportCode: () => {},
+  sortedSubscriptions: [],
+  addSubscription: (
+    label: string,
+    price: number,
+    type: SubscriptionType,
+    date: string,
+  ) => {},
+  calculateTotalPerMonth: () => 0,
+  calculateTotalPerYear: () => 0,
+  handleFormSubmit: (e: FormEvent) => {},
+  handleFormChange: (
+    e: React.ChangeEvent<
+      HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement
+    >,
+  ) => {},
+  editSubscription: () => {},
+  cancelEdit: () => {},
+  deleteSubscription: () => {},
+  getSubscriptionDueDate: () => 0,
+  isSubscriptionDueToday: () => false,
+  getSubscriptionTotalSpend: () => 0,
+  formatToCurrencyString: () => "",
+  backupSubscriptions: async () => {},
+  importSubscriptions: () => {},
+  closeImport: () => {},
+  startImport: () => {},
+  deleteAllSubscriptions: () => {},
+});
+
+export function SubscriptionTrackerProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const [subscriptions, setSubscriptions] = useLocalStorageState<
+    Subscription[]
+  >("subscriptions_tracker", []);
+  const [selectedSubscriptionIndex, setSelectedSubscriptionIndex] =
+    useState<number>(-1);
+  const [formData, setFormData] = useState<Subscription>(DefaultFormData);
+  const [editingSubscription, setEditingSubscription] =
+    useState<boolean>(false);
+  const [locale, setLocale] = useLocalStorageState<string>(
+    "subscription_tracker_locale",
+    "en-UK",
+  );
+  const [currency, setCurrency] = useLocalStorageState<string>(
+    "subscription_tracker_currency",
+    "GBP",
+  );
+  const [isImporting, setIsImporting] = useState<boolean>(false);
+  const [importCode, setImportCode] = useState<string>("");
+
+  const sortedSubscriptions = subscriptions.sort((a, b) => {
+    if (isSubscriptionDueToday(a) !== isSubscriptionDueToday(b)) {
+      return isSubscriptionDueToday(b) ? 1 : -1;
+    }
+    return getSubscriptionDueDate(a) - getSubscriptionDueDate(b);
+  });
+
+  function addSubscription(
+    label: string,
+    price: number,
+    type: SubscriptionType,
+    date: string,
+  ): void {
+    setSubscriptions((prev) => [
+      ...prev,
+      {
+        id: generateUUID(),
+        label: label,
+        price: price,
+        type: type,
+        firstPaymentDate: date,
+      },
+    ]);
+  }
+
+  function calculateTotalPerMonth(): number {
+    return subscriptions
+      .filter((sub) => sub.type === "monthly")
+      .reduce((total, sub) => total + Number(sub.price), 0);
+  }
+
+  function calculateTotalPerYear(): number {
+    const monthlyTotal = calculateTotalPerMonth() * 12;
+    return (
+      monthlyTotal +
+      subscriptions
+        .filter((sub) => sub.type === "yearly")
+        .reduce((total, sub) => total + Number(sub.price), 0)
+    );
+  }
+
+  function handleFormSubmit(e: FormEvent): void {
+    e.preventDefault();
+    if (editingSubscription) {
+      const updatedSubscriptions = subscriptions.map((sub, index) =>
+        index === selectedSubscriptionIndex ? formData : sub,
+      );
+      setSubscriptions(updatedSubscriptions);
+    } else {
+      addSubscription(
+        formData.label,
+        formData.price,
+        formData.type,
+        formData.firstPaymentDate,
+      );
+    }
+    cancelEdit();
+  }
+
+  function handleFormChange(
+    e: React.ChangeEvent<
+      HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement
+    >,
+  ): void {
+    const { name, type, value } = e.target;
+
+    if (type === "checkbox") {
+      const target = e.target as HTMLInputElement;
+      setFormData((prev) => ({
+        ...prev,
+        [name]: target.checked,
+      }));
+    } else {
+      setFormData((prev) => ({
+        ...prev,
+        [name]: value,
+      }));
+    }
+  }
+
+  function editSubscription(index: number): void {
+    if (isIndexOutOfRange(subscriptions, index)) return;
+    const subscription = subscriptions[index];
+    setSelectedSubscriptionIndex(index);
+    setEditingSubscription(true);
+    setFormData({
+      id: subscription.id,
+      label: subscription.label,
+      price: subscription.price,
+      type: subscription.type,
+      firstPaymentDate: subscription.firstPaymentDate,
+    });
+  }
+
+  function cancelEdit(): void {
+    setEditingSubscription(false);
+    setFormData(DefaultFormData);
+  }
+
+  function deleteSubscription(index: number): void {
+    if (isIndexOutOfRange(subscriptions, index)) return;
+    const updatedSubscriptions = subscriptions.filter((_, i) => i !== index);
+    setSubscriptions(updatedSubscriptions);
+    cancelEdit();
+  }
+
+  function getSubscriptionDueDate(subscription: Subscription): number {
+    if (
+      subscription === undefined ||
+      subscription.firstPaymentDate === undefined
+    )
+      return 0;
+    const today = new Date();
+    const nextPaymentDate = new Date(subscription.firstPaymentDate);
+    while (nextPaymentDate <= today) {
+      nextPaymentDate.setMonth(nextPaymentDate.getMonth() + 1);
+    }
+    const timeDiff = nextPaymentDate.getTime() - today.getTime();
+    const days = Math.ceil(timeDiff / (1000 * 60 * 60 * 24));
+    return days;
+  }
+
+  function isSubscriptionDueToday(subscription: Subscription): boolean {
+    if (
+      subscription === undefined ||
+      subscription.firstPaymentDate === undefined
+    )
+      return false;
+    const nextPaymentDate = new Date(subscription.firstPaymentDate);
+    const currentDate = new Date();
+    return nextPaymentDate.toDateString() === currentDate.toDateString();
+  }
+
+  function getSubscriptionTotalSpend(subscription: Subscription): number {
+    if (
+      subscription === undefined ||
+      subscription.firstPaymentDate === undefined
+    )
+      return 0;
+    const today = new Date();
+    const nextPaymentDate = new Date(subscription.firstPaymentDate);
+    let totalSpend = 0;
+    while (nextPaymentDate <= today) {
+      nextPaymentDate.setMonth(nextPaymentDate.getMonth() + 1);
+      totalSpend += Number(subscription.price);
+    }
+    return totalSpend;
+  }
+
+  function formatToCurrencyString(num: number): string {
+    return formatCurrency(num, locale, currency);
+  }
+
+  async function backupSubscriptions(): Promise<void> {
+    const compressedData = compressJSON(subscriptions);
+    await copyToClipboard(compressedData);
+  }
+
+  function importSubscriptions(): void {
+    if (importCode === "") return;
+    const data = decompressJSON(importCode) as Subscription[];
+    setSubscriptions(data);
+    closeImport();
+  }
+
+  function closeImport(): void {
+    setIsImporting(false);
+    setImportCode("");
+  }
+
+  function startImport(): void {
+    setIsImporting(true);
+  }
+
+  function deleteAllSubscriptions(): void {
+    setSubscriptions([]);
+    cancelEdit();
+  }
+
+  return (
+    <SubscriptionTrackerContext.Provider
+      value={{
+        subscriptions,
+        setSubscriptions,
+        selectedSubscriptionIndex,
+        setSelectedSubscriptionIndex,
+        formData,
+        setFormData,
+        editingSubscription,
+        setEditingSubscription,
+        locale,
+        setLocale,
+        currency,
+        setCurrency,
+        isImporting,
+        setIsImporting,
+        importCode,
+        setImportCode,
+        sortedSubscriptions,
+        addSubscription,
+        calculateTotalPerMonth,
+        calculateTotalPerYear,
+        handleFormSubmit,
+        handleFormChange,
+        editSubscription,
+        cancelEdit,
+        deleteSubscription,
+        getSubscriptionDueDate,
+        isSubscriptionDueToday,
+        getSubscriptionTotalSpend,
+        formatToCurrencyString,
+        backupSubscriptions,
+        importSubscriptions,
+        closeImport,
+        startImport,
+        deleteAllSubscriptions,
+      }}
+    >
+      {children}
+    </SubscriptionTrackerContext.Provider>
+  );
+}
+
+export function useSubscriptionTrackerContext() {
+  return useContext(SubscriptionTrackerContext);
+}
