+++ utils/useClampedState.tsx
@@ -1,8 +1,29 @@
-import { useCallback } from "react";
+import { useCallback, useState } from "react";
 import { clamp } from "./helperFunctions";
 import useLocalStorageState from "./useLocalStorageState";
 
-const useClampedState = (
+export const useClampedState = (
+  initialValue: number,
+  min: number,
+  max: number,
+) => {
+  const [value, setValue] = useState(clamp(initialValue, min, max));
+
+  const setClampedValue = useCallback(
+    (newValue: number | ((prevValue: number) => number)) => {
+      setValue((prevValue) => {
+        const nextValue =
+          typeof newValue === "function" ? newValue(prevValue) : newValue;
+        return clamp(nextValue, min, max);
+      });
+    },
+    [min, max],
+  );
+
+  return [value, setClampedValue] as const;
+};
+
+export const useClampedLocalStorageState = (
   key: string,
   initialValue: number,
   min: number,
@@ -26,5 +47,3 @@ const useClampedState = (
 
   return [value, setClampedValue] as const;
 };
-
-export default useClampedState;
